requirements.txt
astroid==2.0.4
autopep8==1.4.2
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
cssselect==1.0.3
dateparser==0.7.0
Django==2.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.5
mccabe==0.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pylint==2.1.1
python-dateutil==2.7.5
pytz==2018.7
regex==2018.11.7
requests==2.20.0
six==1.11.0
tzlocal==1.5.1
urllib3==1.24
wrapt==1.10.11
